requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
pandas>=2.2.0
openpyxl>=3.1.0
groq>=0.4.0
sentence-transformers>=2.2.0
python-docx>=1.0.0
pdfplumber>=0.10.0
